+++ models/relationship.go
@@ -2,6 +2,6 @@ package models
 
 // Relationship Post 和 Meta 的关系表
 type Relationship struct {
-	PostID int `xorm:"pk"`
-	MetaID int `xorm:"pk"`
+	PostID int64 `xorm:"pk"`
+	MetaID int64 `xorm:"pk"`
 }
